+++ examples/day2-examples/src/main/java/com/telekom/training/day2/tarifoptions/initial/BasicTariffWithRoaming.java
@@ -2,10 +2,6 @@ package com.telekom.training.day2.tarifoptions.initial;
 
 import java.math.BigDecimal;
 
-/**
- * PROBLEM: More class explosion
- * What if we want Flatrate AND Roaming? Need another class!
- */
 public class BasicTariffWithRoaming extends BasicTariff {
     private static final BigDecimal ROAMING_COST = new BigDecimal("8.50");